Created a BaseRBM with its extentions
